[Chore] RGB 색상 타입 Int -> UInt8로 변경
