+++ MyKeynote/MyKeynote/Model/Protocols.swift
@@ -1,42 +1,5 @@
 import Foundation
 
-enum AlphaType: Int, CaseIterable {
-    case one = 1
-    case two
-    case three
-    case four
-    case five
-    case six
-    case seven
-    case eight
-    case nine
-    case ten
-    
-    static func randomAlphaType() -> AlphaType {
-        return AlphaType.allCases.randomElement()!
-    }
-}
-
-struct RGBColor {
-    private(set) var red: Int
-    private(set) var green: Int
-    private(set) var blue: Int
-    
-    init(red: Int, green: Int, blue: Int) {
-        self.red = red
-        self.green = green
-        self.blue = blue
-    }
-    
-    static func randomNumberInRange() -> Int {
-        return Int.random(in: 0...256)
-    }
-    
-    static func randomRGBColor() -> RGBColor {
-        return RGBColor(red: randomNumberInRange(), green: randomNumberInRange(), blue: randomNumberInRange())
-    }
-}
-
 protocol Square {
     var rgb: RGBColor { get set}
     var alpha: AlphaType { get set }
@@ -44,12 +7,10 @@ protocol Square {
 }
 
 protocol Slidable {
-    var id: String { get set }
+    var id: String { get }
 }
 
-protocol SquareSlidable: Square, Slidable {
-    
-}
+protocol SquareSlidable: Square, Slidable { }
 
 extension String {
     func subString(from: Int, to: Int) -> String {
